Module vek::geom::repr_simd

Basic geometric primitives that use #[repr(simd)] vectors.

Structs

Aabb
Axis-aligned Bounding Box (3D), represented by min and max points.

Aabr
Axis-aligned Bounding Rectangle (2D), represented by min and max points.

Disk
Disk (2D), represented by center and radius.

Ellipsis
Ellipsis (2D), represented by center and radius in both axii.

LineSegment2
2D Line segment, represented by two points, start and end.

LineSegment3
3D Line segment, represented by two points, start and end.

Potato
Nobody can possibly use this ???

Ray
3D ray, represented by a starting point and a normalized direction vector.

Rect
2D rectangle, represented by a bottom-left position, and 2D extents.

Rect3
A Rect extended to 3D.

Sphere
Sphere (3D), represented by center and radius.