Module vek::mat::repr_simd::row_major

Matrices stored in row-major layout.

Unlike the column-major layout, row-major matrices are good at being the right-hand-side of a product with one or more row vectors.

Also, their natural indexing order matches the existing mathematical conventions.

Re-exports

pub use self::mat4::Mat4;

pub use self::mat3::Mat3;

pub use self::mat2::Mat2;

Modules

mat2
2x2 matrix.

mat3
3x3 matrix.

mat4
4x4 matrix.