Struct vek::mat::repr_simd::row_major::mat2::Mat2

#[repr(C)]
pub struct Mat2<T> {
    pub rows: CVec2<Vec2<T>>,
}

2x2 matrix.

Fields
rows: CVec2<Vec2<T>>
Implementations

impl<T> Mat2<T>

pub fn identity() -> Self where
T: Zero + One,

The identity matrix, which is also the default value for square matrices.

assert_eq!(Mat4::<f32>::default(), Mat4::<f32>::identity());

pub fn zero() -> Self where
T: Zero,

The matrix with all elements set to zero.

pub fn apply<F>(&mut self, f: F) where
T: Copy,
F: FnMut(T) -> T,

Applies the function f to each element of this matrix, in-place.

For an example, see the map() method.

pub fn apply2<F, S>(&mut self, other: Mat2<S>, f: F) where
T: Copy,
F: FnMut(T, S) -> T,

Applies the function f to each element of this matrix, in-place.

For an example, see the map2() method.

pub fn numcast<D>(self) -> Option<Mat2<D>> where
T: NumCast,
D: NumCast,

Returns a memberwise-converted copy of this matrix, using NumCast.

let m = Mat4::<f32>::identity();
let m: Mat4<i32> = m.numcast().unwrap();
assert_eq!(m, Mat4::identity());

pub fn broadcast_diagonal(val: T) -> Self where
T: Zero + Copy,

Initializes a new matrix with elements of the diagonal set to val and the other to zero.

In a way, this is the same as single-argument matrix constructors in GLSL and GLM.

assert_eq!(Mat4::broadcast_diagonal(0), Mat4::zero());
assert_eq!(Mat4::broadcast_diagonal(1), Mat4::identity());
assert_eq!(Mat4::broadcast_diagonal(2), Mat4::new(
    2,0,0,0,
    0,2,0,0,
    0,0,2,0,
    0,0,0,2,
));

pub fn with_diagonal(d: Vec2<T>) -> Self where
T: Zero + Copy,

Initializes a matrix by its diagonal, setting other elements to zero.

pub fn diagonal(self) -> Vec2<T>

Gets the matrix’s diagonal into a vector.

assert_eq!(Mat4::<u32>::zero().diagonal(), Vec4::zero());
assert_eq!(Mat4::<u32>::identity().diagonal(), Vec4::one());

let mut m = Mat4::zero();
m[(0, 0)] = 1;
m[(1, 1)] = 2;
m[(2, 2)] = 3;
m[(3, 3)] = 4;
assert_eq!(m.diagonal(), Vec4::new(1, 2, 3, 4));
assert_eq!(m.diagonal(), Vec4::iota() + 1);

pub fn trace(self) -> T where
T: Add<T, Output = T>,

The sum of the diagonal’s elements.

assert_eq!(Mat4::<u32>::zero().trace(), 0);
assert_eq!(Mat4::<u32>::identity().trace(), 4);

pub fn mul_memberwise(self, m: Self) -> Self where
T: Mul<Output = T>,

Multiply elements of this matrix with another’s.

let m = Mat4::new(
    0, 1, 2, 3,
    1, 2, 3, 4,
    2, 3, 4, 5,
    3, 4, 5, 6,
);
let r = Mat4::new(
    0, 1, 4, 9,
    1, 4, 9, 16,
    4, 9, 16, 25,
    9, 16, 25, 36,
);
assert_eq!(m.mul_memberwise(m), r);

pub fn row_count(&self) -> usize

Convenience for getting the number of rows of this matrix.

pub fn col_count(&self) -> usize

Convenience for getting the number of columns of this matrix.

pub const ROW_COUNT: usize

Convenience constant representing the number of rows for matrices of this type.

pub const COL_COUNT: usize

Convenience constant representing the number of columns for matrices of this type.

pub fn is_packed(&self) -> bool

Are all elements of this matrix tightly packed together in memory ?

This might not be the case for matrices in the repr_simd module (it depends on the target architecture).

impl<T> Mat2<T>

pub fn map_rows<D, F>(self, f: F) -> Mat2<D> where
F: FnMut(Vec2<T>) -> Vec2<D>,

Returns a row-wise-converted copy of this matrix, using the given conversion closure.

use vek::mat::repr_c::row_major::Mat4;

let m = Mat4::<f32>::new(
    0.25, 1.25, 5.56, 8.66,
    8.53, 2.92, 3.86, 9.36,
    1.02, 0.28, 5.52, 6.06,
    6.20, 7.01, 4.90, 5.26
);
let m = m.map_rows(|row| row.map(|x| x.round() as i32));
assert_eq!(m, Mat4::new(
    0, 1, 6, 9,
    9, 3, 4, 9,
    1, 0, 6, 6,
    6, 7, 5, 5
));

pub fn into_row_array(self) -> [T; 4]

Converts this matrix into a fixed-size array of elements.

use vek::mat::repr_c::row_major::Mat4;

let m = Mat4::<u32>::new(
     0,  1,  2,  3,
     4,  5,  6,  7,
     8,  9, 10, 11,
    12, 13, 14, 15
);
let array = [
     0,  1,  2,  3,
     4,  5,  6,  7,
     8,  9, 10, 11,
    12, 13, 14, 15
];
assert_eq!(m.into_row_array(), array);

pub fn into_row_arrays(self) -> [[T; 2]; 2]

Converts this matrix into a fixed-size array of fixed-size arrays of elements.

use vek::mat::repr_c::row_major::Mat4;

let m = Mat4::<u32>::new(
     0,  1,  2,  3,
     4,  5,  6,  7,
     8,  9, 10, 11,
    12, 13, 14, 15
);
let array = [
    [  0,  1,  2,  3, ],
    [  4,  5,  6,  7, ],
    [  8,  9, 10, 11, ],
    [ 12, 13, 14, 15, ],
];
assert_eq!(m.into_row_arrays(), array);

pub fn from_row_array(array: [T; 4]) -> Self

Converts a fixed-size array of elements into a matrix.

use vek::mat::repr_c::row_major::Mat4;

let m = Mat4::<u32>::new(
     0,  1,  2,  3,
     4,  5,  6,  7,
     8,  9, 10, 11,
    12, 13, 14, 15
);
let array = [
     0,  1,  2,  3,
     4,  5,  6,  7,
     8,  9, 10, 11,
    12, 13, 14, 15
];
assert_eq!(m, Mat4::from_row_array(array));

pub fn from_row_arrays(array: [[T; 2]; 2]) -> Self

Converts a fixed-size array of fixed-size array of elements into a matrix.

use vek::mat::repr_c::row_major::Mat4;

let m = Mat4::<u32>::new(
     0,  1,  2,  3,
     4,  5,  6,  7,
     8,  9, 10, 11,
    12, 13, 14, 15
);
let array = [
    [  0,  1,  2,  3, ],
    [  4,  5,  6,  7, ],
    [  8,  9, 10, 11, ],
    [ 12, 13, 14, 15, ],
];
assert_eq!(m, Mat4::from_row_arrays(array));

pub fn into_col_array(self) -> [T; 4]

Converts this matrix into a fixed-size array of elements.

use vek::mat::repr_c::row_major::Mat4;

let m = Mat4::<u32>::new(
     0,  1,  2,  3,
     4,  5,  6,  7,
     8,  9, 10, 11,
    12, 13, 14, 15
);
let array = [
    0, 4, 8, 12,
    1, 5, 9, 13,
    2, 6, 10, 14,
    3, 7, 11, 15
];
assert_eq!(m.into_col_array(), array);

pub fn into_col_arrays(self) -> [[T; 2]; 2]

Converts this matrix into a fixed-size array of fixed-size arrays of elements.

use vek::mat::repr_c::row_major::Mat4;

let m = Mat4::<u32>::new(
     0,  1,  2,  3,
     4,  5,  6,  7,
     8,  9, 10, 11,
    12, 13, 14, 15
);
let array = [
    [ 0, 4,  8, 12, ],
    [ 1, 5,  9, 13, ],
    [ 2, 6, 10, 14, ],
    [ 3, 7, 11, 15, ],
];
assert_eq!(m.into_col_arrays(), array);

pub fn from_col_array(array: [T; 4]) -> Self

Converts a fixed-size array of elements into a matrix.

use vek::mat::repr_c::row_major::Mat4;

let m = Mat4::<u32>::new(
     0,  1,  2,  3,
     4,  5,  6,  7,
     8,  9, 10, 11,
    12, 13, 14, 15
);
let array = [
    0, 4, 8, 12,
    1, 5, 9, 13,
    2, 6, 10, 14,
    3, 7, 11, 15
];
assert_eq!(m, Mat4::from_col_array(array));

pub fn from_col_arrays(array: [[T; 2]; 2]) -> Self

Converts a fixed-size array of fixed-size arrays of elements into a matrix.

use vek::mat::repr_c::row_major::Mat4;

let m = Mat4::<u32>::new(
     0,  1,  2,  3,
     4,  5,  6,  7,
     8,  9, 10, 11,
    12, 13, 14, 15
);
let array = [
    [ 0, 4,  8, 12, ],
    [ 1, 5,  9, 13, ],
    [ 2, 6, 10, 14, ],
    [ 3, 7, 11, 15, ],
];
assert_eq!(m, Mat4::from_col_arrays(array));

pub fn as_row_ptr(&self) -> *const T

Gets a const pointer to this matrix’s elements.

Panics

Panics if the matrix’s elements are not tightly packed in memory, which may be the case for matrices in the repr_simd module. You may check this with the is_packed() method.

pub fn as_mut_row_ptr(&mut self) -> *mut T

Gets a mut pointer to this matrix’s elements.

Panics

Panics if the matrix’s elements are not tightly packed in memory, which may be the case for matrices in the repr_simd module. You may check this with the is_packed() method.

pub fn as_row_slice(&self) -> &[T]

View this matrix as an immutable slice.

Panics

Panics if the matrix’s elements are not tightly packed in memory, which may be the case for matrices in the repr_simd module. You may check this with the is_packed() method.

pub fn as_mut_row_slice(&mut self) -> &mut [T]

View this matrix as a mutable slice.

Panics

Panics if the matrix’s elements are not tightly packed in memory, which may be the case for matrices in the repr_simd module. You may check this with the is_packed() method.

impl<T> Mat2<T>

pub fn gl_should_transpose(&self) -> bool

Gets the transpose parameter to pass to OpenGL glUniformMatrix*() functions.

The return value is a plain bool which you may directly cast to a GLboolean.

This takes &self to prevent surprises when changing the type of matrix you plan to send.

pub const GL_SHOULD_TRANSPOSE: bool

The transpose parameter to pass to OpenGL glUniformMatrix*() functions.

impl<T> Mat2<T>

pub fn map<D, F>(self, f: F) -> Mat2<D> where
F: FnMut(T) -> D,

Returns an element-wise-converted copy of this matrix, using the given conversion closure.

use vek::mat::repr_c::row_major::Mat4;

let m = Mat4::<f32>::new(
    0.25, 1.25, 5.56, 8.66,
    8.53, 2.92, 3.86, 9.36,
    1.02, 0.28, 5.52, 6.06,
    6.20, 7.01, 4.90, 5.26
);
let m = m.map(|x| x.round() as i32);
assert_eq!(m, Mat4::new(
    0, 1, 6, 9,
    9, 3, 4, 9,
    1, 0, 6, 6,
    6, 7, 5, 5
));

pub fn as_<D>(self) -> Mat2<D> where
T: AsPrimitive<D>,
D: 'static + Copy,

Returns a memberwise-converted copy of this matrix, using AsPrimitive.

Examples

let v = Vec4::new(0_f32, 1., 2., 3.);
let i: Vec4<i32> = v.as_();
assert_eq!(i, Vec4::new(0, 1, 2, 3));

Safety

In Rust versions before 1.45.0, some uses of the as operator were not entirely safe. In particular, it was undefined behavior if a truncated floating point value could not fit in the target integer type (#10184);

let x: u8 = (1.04E+17).as_(); // UB

pub fn map2<D, F, S>(self, other: Mat2<S>, f: F) -> Mat2<D> where
F: FnMut(T, S) -> D,

Applies the function f to each element of two matrices, pairwise, and returns the result.

use vek::mat::repr_c::row_major::Mat4;

let a = Mat4::<f32>::new(
    0.25, 1.25, 2.52, 2.99,
    0.25, 1.25, 2.52, 2.99,
    0.25, 1.25, 2.52, 2.99,
    0.25, 1.25, 2.52, 2.99
);
let b = Mat4::<i32>::new(
    0, 1, 0, 0,
    1, 0, 0, 0,
    0, 0, 1, 0,
    0, 0, 0, 1
);
let m = a.map2(b, |a, b| a.round() as i32 + b);
assert_eq!(m, Mat4::new(
    0, 2, 3, 3,
    1, 1, 3, 3,
    0, 1, 4, 3,
    0, 1, 3, 4
));

pub fn transposed(self) -> Self

The matrix’s transpose.

let m = Mat2::new(
    0, 1,
    4, 5
);
let t = Mat2::new(
    0, 4,
    1, 5
);
assert_eq!(m.transposed(), t);
assert_eq!(m, m.transposed().transposed());

pub fn transpose(&mut self)

Transpose this matrix.

let mut m = Mat2::new(
    0, 1,
    4, 5
);
let t = Mat2::new(
    0, 4,
    1, 5
);
m.transpose();
assert_eq!(m, t);

pub fn determinant(self) -> T where
T: Mul<T, Output = T> + Sub<T, Output = T>,

Get this matrix’s determinant.

A matrix is invertible if its determinant is non-zero.

let (a,b,c,d) = (0,1,2,3);
let m = Mat2::new(a,b,c,d);
assert_eq!(m.determinant(), a*d - c*b);

pub fn rotate_z(&mut self, angle_radians: T) where
T: Real + MulAdd<T, T, Output = T>,

Rotates this matrix around the Z axis (counter-clockwise rotation in 2D).

pub fn rotated_z(self, angle_radians: T) -> Self where
T: Real + MulAdd<T, T, Output = T>,

Rotates this matrix around the Z axis (counter-clockwise rotation in 2D).

pub fn rotation_z(angle_radians: T) -> Self where
T: Real,

Creates a matrix that rotates around the Z axis (counter-clockwise rotation in 2D).

pub fn scale_2d<V: Into<Vec2<T>>>(&mut self, v: V) where
T: Real + MulAdd<T, T, Output = T>,

Scales this matrix in 2D.

pub fn scaled_2d<V: Into<Vec2<T>>>(self, v: V) -> Self where
T: Real + MulAdd<T, T, Output = T>,

Returns this matrix scaled in 2D.

pub fn scaling_2d<V: Into<Vec2<T>>>(v: V) -> Self where
T: Zero,

Creates a 2D scaling matrix.

pub fn shear_x(&mut self, k: T) where
T: Real + MulAdd<T, T, Output = T>,

Shears this matrix along the X axis.

pub fn sheared_x(self, k: T) -> Self where
T: Real + MulAdd<T, T, Output = T>,

Returns this matrix sheared along the X axis.

pub fn shearing_x(k: T) -> Self where
T: Zero + One,

Creates a 2D shearing matrix along the X axis.

pub fn shear_y(&mut self, k: T) where
T: Real + MulAdd<T, T, Output = T>,

Shears this matrix along the Y axis.

pub fn sheared_y(self, k: T) -> Self where
T: Real + MulAdd<T, T, Output = T>,

Returns this matrix sheared along the Y axis.

pub fn shearing_y(k: T) -> Self where
T: Zero + One,

Creates a 2D shearing matrix along the Y axis.

impl<T> Mat2<T>

pub fn new(m00: T, m01: T, m10: T, m11: T) -> Self

Creates a new 2x2 matrix from elements in a layout-agnostic way.

The parameters are named mij where i is the row index and j the column index. Their order is always the same regardless of the matrix’s layout.