Module vek::ops

Operations defined by this crate, such as MulAdd, Lerp, Clamp, and Wrap.

Traits

Clamp
A scalar or vector that can be constrained to be between two values (inclusive).

Clamp01
A scalar or vector that can be constrained to be between 0 and 1 (inclusive).

ClampMinus1
A scalar or vector that can be constrained to be between -1 and 1 (inclusive).

ColorComponent
Trait for types that are suitable for representing a color component value.

IsBetween
A value that can tell whether or not it is between two bounds (inclusive).

IsBetween01
A value that can tell whether or not it is between 0 and 1 (inclusive).

Lerp
A value that can be linearly interpolated.

MulAdd
Fused multiply-add. Computes (self * a) + b with only one rounding error, yielding a more accurate result than an unfused multiply-add.

Slerp
A value that can be Spherically Linearly interpolated.

Wrap
A value that can wrap itself around given bounds.

Functions

mul_add	Deprecated
Returns (val * mul) + add.

partial_max
Compares and returns the maximum of two values, using partial ordering.

partial_min
Compares and returns the minimum of two values, using partial ordering.