Trait vek::ops::IsBetween

pub trait IsBetween<Bound = Self>: Sized {
    type Output;
    fn is_between(self, lower: Bound, upper: Bound) -> Self::Output;

    fn is_between01(self) -> Self::Output
    where
        Bound: Zero + One,
    { ... }
}

A value that can tell whether or not it is between two bounds (inclusive).

Associated Types

type Output

bool for scalars, or vector of bools for vectors.

Loading content...
Required methods

fn is_between(self, lower: Bound, upper: Bound) -> Self::Output

Returns whether this value is between lower and upper (inclusive).

This would rather make use of inclusive ranges, but it’s an unstable feature.

Panics

Panics if lower is greater than upper. Swap the values yourself if necessary.

use vek::ops::IsBetween;

assert!(5_i32 .is_between(5, 10));
assert!(7_i32 .is_between(5, 10));
assert!(10_i32.is_between(5, 10));
assert!(!(4_i32 .is_between(5, 10)));
assert!(!(11_i32.is_between(5, 10)));

Loading content...
Provided methods

fn is_between01(self) -> Self::Output where
Bound: Zero + One,

Returns whether this value is between 0 and 1 (inclusive).

Loading content...