Trait vek::ops::Slerp

pub trait Slerp<Factor = f32>: Sized {
    type Output;
    fn slerp_unclamped(from: Self, to: Self, factor: Factor) -> Self::Output;

    fn slerp(from: Self, to: Self, factor: Factor) -> Self::Output
    where
        Factor: Clamp + Zero + One,
    { ... }
}

A value that can be Spherically Linearly interpolated.

The Output type allows this trait to be meaningfully implemented for &T as well as T.

Associated Types

type Output

The resulting type after performing the SLERP operation.

Loading content...
Required methods

fn slerp_unclamped(from: Self, to: Self, factor: Factor) -> Self::Output

Performs spherical linear interpolation without implictly constraining factor to be between 0 and 1.

Loading content...
Provided methods

fn slerp(from: Self, to: Self, factor: Factor) -> Self::Output where
Factor: Clamp + Zero + One,

Performs spherical linear interpolation, constraining factor to be between 0 and 1.

Loading content...