Struct vek::quaternion::repr_c::Quaternion

#[repr(C)]
pub struct Quaternion<T> {
    pub x: T,
    pub y: T,
    pub z: T,
    pub w: T,
}

Quaternions are a convenient representation for rotations in 3D spaces.

IMPORTANT: Quaternions are only valid as rotations as long as they are normalized (i.e their magnitude is 1). Most operations assume this, instead of normalizing inputs behind your back, so be careful.

They essentially consist of a vector part (x, y, z), and scalar part (w). For unit quaternions, the vector part is the unit axis of rotation scaled by the sine of the half-angle of the rotation, and the scalar part is the cosine of the half-angle.

Fields
x: Ty: Tz: Tw: T
Implementations

impl<T> Quaternion<T>

pub fn from_xyzw(x: T, y: T, z: T, w: T) -> Self

Creates a new quaternion with x, y, z and w elements in order.

You are responsible for ensuring that the resulting quaternion is normalized.

pub fn from_scalar_and_vec3<V: Into<Vec3<T>>>(pair: (T, V)) -> Self

Creates a new quaternion from a scalar-and-vector pair.

You are responsible for ensuring that the resulting quaternion is normalized.

pub fn into_scalar_and_vec3(self) -> (T, Vec3<T>)

Converts this quaternion into a scalar-and-vector pair by destructuring.

Not to be confused with into_angle_axis().

pub fn zero() -> Self where
T: Zero,

Creates a new quaternion with all elements set to zero.

Be careful: since it has a magnitude equal to zero, it is not valid to use for most operations.

pub fn identity() -> Self where
T: Zero + One,

Creates the identity quaternion.

use std::f32::consts::PI;

let id = Quaternion::<f32>::identity();
assert_eq!(id, Default::default());
assert_relative_eq!(id, id.conjugate());
assert_relative_eq!(id, id.inverse());

let q = Quaternion::rotation_y(PI);
assert_relative_eq!(id * q, q);
assert_relative_eq!(q * id, q);

pub fn conjugate(self) -> Self where
T: Neg<Output = T>,

Gets this quaternion’s conjugate (copy with negated vector part).

On normalized quaternions, the conjugate also happens to be the inverse.

use std::f32::consts::PI;

let p = Quaternion::rotation_x(PI);
let q = Quaternion::rotation_z(PI);
assert_relative_eq!((p*q).conjugate(), q.conjugate() * p.conjugate());

// Rotation quaternions are normalized, so their conjugate is also their inverse.
assert_relative_eq!(q.conjugate(), q.inverse());

pub fn inverse(self) -> Self where
T: Neg<Output = T> + Copy + Add<T, Output = T> + Mul<Output = T> + Div<Output = T>,

Gets this quaternion’s inverse, i.e the one that reverses its effect.

On normalized quaternions, the inverse happens to be the conjugate.

use std::f32::consts::PI;

let rot = Quaternion::rotation_y(PI);
let inv = rot.inverse();
assert_relative_eq!(rot*inv, Quaternion::identity());
assert_relative_eq!(inv*rot, Quaternion::identity());

let p = Quaternion::rotation_x(PI);
let q = Quaternion::rotation_z(PI);
assert_relative_eq!((p*q).inverse(), q.inverse() * p.inverse());

pub fn dot(self, q: Self) -> T where
T: Copy + Add<T, Output = T> + Mul<Output = T>,

Gets the dot product between two quaternions.

pub fn normalized(self) -> Self where
T: Real + Add<T, Output = T>,

Gets a normalized copy of this quaternion.

pub fn magnitude_squared(self) -> T where
T: Real + Add<T, Output = T>,

Gets this quaternion’s magnitude, squared.

pub fn magnitude(self) -> T where
T: Real + Add<T, Output = T>,

Gets this quaternion’s magnitude.

pub fn rotation_from_to_3d<V: Into<Vec3<T>>>(from: V, to: V) -> Self where
T: Real + Add<T, Output = T>,

Creates a quaternion that would rotate a from direction to to.

let (from, to) = (Vec4::<f32>::unit_x(), Vec4::<f32>::unit_y());
let q = Quaternion::<f32>::rotation_from_to_3d(from, to);
assert_relative_eq!(q * from, to);
assert_relative_eq!(q * Vec4::unit_y(), -Vec4::unit_x());

let (from, to) = (Vec4::<f32>::unit_x(), -Vec4::<f32>::unit_x());
let q = Quaternion::<f32>::rotation_from_to_3d(from, to);
assert_relative_eq!(q * from, to);

pub fn rotation_3d<V: Into<Vec3<T>>>(angle_radians: T, axis: V) -> Self where
T: Real + Add<T, Output = T>,

Creates a quaternion from an angle and axis. The axis is not required to be normalized.

pub fn rotation_x(angle_radians: T) -> Self where
T: Real + Add<T, Output = T>,

Creates a quaternion from an angle for a rotation around the X axis.

pub fn rotation_y(angle_radians: T) -> Self where
T: Real + Add<T, Output = T>,

Creates a quaternion from an angle for a rotation around the Y axis.

pub fn rotation_z(angle_radians: T) -> Self where
T: Real + Add<T, Output = T>,

Creates a quaternion from an angle for a rotation around the Y axis.

pub fn rotated_3d<V: Into<Vec3<T>>>(self, angle_radians: T, axis: V) -> Self where
T: Real + Add<T, Output = T>,

Returns this quaternion rotated around the given axis with given angle. The axis is not required to be normalized.

pub fn rotated_x(self, angle_radians: T) -> Self where
T: Real + Add<T, Output = T>,

Returns this quaternion rotated around the X axis with given angle.

pub fn rotated_y(self, angle_radians: T) -> Self where
T: Real + Add<T, Output = T>,

Returns this quaternion rotated around the Y axis with given angle.

pub fn rotated_z(self, angle_radians: T) -> Self where
T: Real + Add<T, Output = T>,

Returns this quaternion rotated around the Z axis with given angle.

pub fn rotate_3d<V: Into<Vec3<T>>>(&mut self, angle_radians: T, axis: V) where
T: Real + Add<T, Output = T>,

Rotates this quaternion around the given axis with given angle. The axis is not required to be normalized.

pub fn rotate_x(&mut self, angle_radians: T) where
T: Real + Add<T, Output = T>,

Rotates this quaternion around the X axis with given angle.

pub fn rotate_y(&mut self, angle_radians: T) where
T: Real + Add<T, Output = T>,

Rotates this quaternion around the Y axis with given angle.

pub fn rotate_z(&mut self, angle_radians: T) where
T: Real + Add<T, Output = T>,

Rotates this quaternion around the Z axis with given angle.

pub fn into_angle_axis(self) -> (T, Vec3<T>) where
T: Real,

Convert this quaternion to angle-axis representation, assuming the quaternion is normalized.

use std::f32::consts::PI;

let q = Quaternion::rotation_x(PI/2.);
let (angle, axis) = q.into_angle_axis();
assert_relative_eq!(angle, PI/2.);
assert_relative_eq!(axis, Vec3::unit_x());

let angle = PI*4./5.;
let axis = Vec3::new(1_f32, 2., 3.);
let q = Quaternion::rotation_3d(angle, axis);
let (a, v) = q.into_angle_axis();
assert_relative_eq!(a, angle);
assert_relative_eq!(v, axis.normalized());

pub fn into_vec4(self) -> Vec4<T>

Converts this quaternion to a Vec4 by destructuring.

pub fn from_vec4(v: Vec4<T>) -> Self

Creates a quaternion from a Vec4 by destructuring. You are responsible for ensuring that the resulting quaternion is normalized.

pub fn into_vec3(self) -> Vec3<T>

Converts this quaternion to a Vec3 by destructuring, dropping the w element.

impl<T> Quaternion<T> where
T: Copy + One + Mul<Output = T> + Sub<Output = T> + MulAdd<T, T, Output = T>,

pub fn lerp_precise_unnormalized(from: Self, to: Self, factor: T) -> Self where
T: Clamp + Zero,

Performs linear interpolation without normalizing the result, using an implementation that supposedly yields a more precise result.

This is probably not what you’re looking for. For an implementation that normalizes the result (which is more commonly wanted), see the Lerp implementation.

pub fn lerp_unclamped_precise_unnormalized(
from: Self,
to: Self,
factor: T
) -> Self

Performs linear interpolation without normalizing the result and without implicitly constraining factor to be between 0 and 1, using an implementation that supposedly yields a more precise result.

This is probably not what you’re looking for. For an implementation that normalizes the result (which is more commonly wanted), see the Lerp implementation.

impl<T> Quaternion<T> where
T: Copy + Sub<Output = T> + MulAdd<T, T, Output = T>,

pub fn lerp_unnormalized(from: Self, to: Self, factor: T) -> Self where
T: Clamp + Zero + One,

Performs linear interpolation without normalizing the result.

This is probably not what you’re looking for. For an implementation that normalizes the result (which is more commonly wanted), see the Lerp implementation.

pub fn lerp_unclamped_unnormalized(from: Self, to: Self, factor: T) -> Self

Performs linear interpolation without normalizing the result and without implicitly constraining factor to be between 0 and 1.

This is probably not what you’re looking for. For an implementation that normalizes the result (which is more commonly wanted), see the Lerp implementation.

impl<T> Quaternion<T> where
T: Lerp<T, Output = T> + Add<T, Output = T> + Real,

pub fn slerp_unclamped(from: Self, to: Self, factor: T) -> Self

Performs spherical linear interpolation without implictly constraining factor to be between 0 and 1.

use std::f32::consts::PI;

let from = Quaternion::rotation_z(0_f32);
let to = Quaternion::rotation_z(PI*9./10.);

let angles = 32;
for i in 0..angles {
    let factor = (i as f32) / (angles as f32);
    let expected = Quaternion::rotation_z(factor * PI*9./10.);
    let slerp = Quaternion::slerp(from, to, factor);
    assert_relative_eq!(slerp, expected);
}

pub fn slerp(from: Self, to: Self, factor: T) -> Self where
T: Clamp,

Perform spherical linear interpolation, constraining factor to be between 0 and 1.