Struct vek::transform::repr_simd::Transform

pub struct Transform<P, O, S> {
    pub position: Vec3<P>,
    pub orientation: Quaternion<O>,
    pub scale: Vec3<S>,
}

A convenient position + orientation + scale container, backed by two Vec3 and a Quaternion.

It can be easily interpolated and converted to a Mat4 of any layout.

let (p, rz, s) = (Vec3::unit_x(), 3.0_f32, 5.0_f32);
let a = Mat4::scaling_3d(s).rotated_z(rz).translated_3d(p);
let b = Mat4::from(Transform {
    position: p, 
    orientation: Quaternion::rotation_z(rz),
    scale: Vec3::broadcast(s),
});
assert_relative_eq!(a, b);

Fields
position: Vec3<P>

Local position.

orientation: Quaternion<O>

Local orientation; It is not named rotation because rotation denotes an operation, but not a current state.

scale: Vec3<S>

Local scale.