Trait vek::transition::ProgressMapper

pub trait ProgressMapper<Progress = f32> {
    fn map_progress(&self, progress: Progress) -> Progress;
}

A functor that maps a progress value to a LERP factor.

Required methods

fn map_progress(&self, progress: Progress) -> Progress

Maps an unconstrained progress value to a LERP factor.

The input progress value should be between zero and one, but is not required to.
The returned LERP factor should be between zero and one, but is not required to.

Loading content...