Struct vek::vec::repr_c::vec3::Vec3

#[repr(C)]
pub struct Vec3<T> {
    pub x: T,
    pub y: T,
    pub z: T,
}

Vector type suited for 3D spatial coordinates.

Fields
x: Ty: Tz: T
Implementations

impl<T> Vec3<T>

pub const fn new(x: T, y: T, z: T) -> Self

Creates a vector from elements.

impl<T> Vec3<T>

pub fn broadcast(val: T) -> Self where
T: Copy,

Broadcasts a single value to all elements of a new vector.

This function is also named splat() in some libraries, or set1() in Intel intrinsics.

“Broadcast” was chosen as the name because it is explicit enough and is the same wording as the description in relevant Intel intrinsics.

assert_eq!(Vec4::broadcast(5), Vec4::new(5,5,5,5));
assert_eq!(Vec4::broadcast(5), Vec4::from(5));

pub fn zero() -> Self where
T: Zero,

Creates a new vector with all elements set to zero.

assert_eq!(Vec4::zero(), Vec4::new(0,0,0,0));
assert_eq!(Vec4::zero(), Vec4::broadcast(0));
assert_eq!(Vec4::zero(), Vec4::from(0));

pub fn one() -> Self where
T: One,

Creates a new vector with all elements set to one.

assert_eq!(Vec4::one(), Vec4::new(1,1,1,1));
assert_eq!(Vec4::one(), Vec4::broadcast(1));
assert_eq!(Vec4::one(), Vec4::from(1));

pub fn iota() -> Self where
T: Zero + One + AddAssign + Copy,

Produces a vector of the first n integers, starting from zero, where n is the number of elements for this vector type.

The iota (ι) function, originating from APL.

See this StackOverflow answer.

This is mostly useful for debugging purposes and tests.

assert_eq!(Vec4::iota(), Vec4::new(0, 1, 2, 3));

pub const fn elem_count(&self) -> usize

Convenience method which returns the number of elements of this vector.

let v = Vec4::new(0,1,2,3);
assert_eq!(v.elem_count(), 4);

pub const ELEM_COUNT: usize

Convenience constant representing the number of elements for this vector type.

pub fn into_tuple(self) -> (T, T, T)

Converts this into a tuple with the same number of elements by consuming.

pub fn into_array(self) -> [T; 3]

Converts this vector into a fixed-size array.

pub fn as_slice(&self) -> &[T]

View this vector as an immutable slice.

pub fn as_mut_slice(&mut self) -> &mut [T]

View this vector as a mutable slice.

pub fn from_slice(slice: &[T]) -> Self where
T: Default + Copy,

Collects the content of a slice into a new vector. Elements are initialized to their default values.

pub fn map<D, F>(self, f: F) -> Vec3<D> where
F: FnMut(T) -> D,

Returns a memberwise-converted copy of this vector, using the given conversion closure.

let v = Vec4::new(0_f32, 1., 1.8, 3.14);
let i = v.map(|x| x.round() as i32);
assert_eq!(i, Vec4::new(0, 1, 2, 3));

Performing LERP on integer vectors by concisely converting them to floats:

let a = Vec4::new(0,1,2,3).map(|x| x as f32);
let b = Vec4::new(2,3,4,5).map(|x| x as f32);
let v = Vec4::lerp(a, b, 0.5_f32).map(|x| x.round() as i32);
assert_eq!(v, Vec4::new(1,2,3,4));

pub fn map2<D, F, S>(self, other: Vec3<S>, f: F) -> Vec3<D> where
F: FnMut(T, S) -> D,

Applies the function f to each element of two vectors, pairwise, and returns the result.

let a = Vec4::<u8>::new(255, 254, 253, 252);
let b = Vec4::<u8>::new(1, 2, 3, 4);
let v = a.map2(b, |a, b| a.wrapping_add(b));
assert_eq!(v, Vec4::zero());
let v = a.map2(b, u8::wrapping_add);
assert_eq!(v, Vec4::zero());

pub fn map3<D, F, S1, S2>(self, a: Vec3<S1>, b: Vec3<S2>, f: F) -> Vec3<D> where
F: FnMut(T, S1, S2) -> D,

Applies the function f to each element of three vectors, and returns the result.

let a = Vec4::<u8>::new(255, 254, 253, 252);
let b = Vec4::<u8>::new(1, 2, 3, 4);
let c = Vec4::<u8>::new(1, 2, 3, 4);
let v = a.map3(b, c, |a, b, c| a.wrapping_add(b) + c);
assert_eq!(v, c);

pub fn apply<F>(&mut self, f: F) where
T: Copy,
F: FnMut(T) -> T,

Applies the function f to each element of this vector, in-place.

let mut v = Vec4::new(0_u32, 1, 2, 3);
v.apply(|x| x.count_ones());
assert_eq!(v, Vec4::new(0, 1, 1, 2));

pub fn apply2<F, S>(&mut self, other: Vec3<S>, f: F) where
T: Copy,
F: FnMut(T, S) -> T,

Applies the function f to each element of two vectors, pairwise, in-place.

let mut a = Vec4::<u8>::new(255, 254, 253, 252);
let b = Vec4::<u8>::new(1, 2, 3, 4);
a.apply2(b, |a, b| a.wrapping_add(b));
assert_eq!(a, Vec4::zero());
a.apply2(b, u8::wrapping_add);
assert_eq!(a, b);

pub fn apply3<F, S1, S2>(&mut self, a: Vec3<S1>, b: Vec3<S2>, f: F) where
T: Copy,
F: FnMut(T, S1, S2) -> T,

Applies the function f to each element of three vectors, in-place.

let mut a = Vec4::<u8>::new(255, 254, 253, 252);
let b = Vec4::<u8>::new(1, 2, 3, 4);
let c = Vec4::<u8>::new(1, 2, 3, 4);
a.apply3(b, c, |a, b, c| a.wrapping_add(b) + c);
assert_eq!(a, c);

pub fn zip<S>(self, other: Vec3<S>) -> Vec3<(T, S)>

“Zips” two vectors together into a vector of tuples.

let a = Vec4::<u8>::new(255, 254, 253, 252);
let b = Vec4::<u8>::new(1, 2, 3, 4);
assert_eq!(a.zip(b), Vec4::new((255, 1), (254, 2), (253, 3), (252, 4)));

pub fn as_<D>(self) -> Vec3<D> where
T: AsPrimitive<D>,
D: 'static + Copy,

Returns a memberwise-converted copy of this vector, using AsPrimitive.

Examples

let v = Vec4::new(0_f32, 1., 2., 3.);
let i: Vec4<i32> = v.as_();
assert_eq!(i, Vec4::new(0, 1, 2, 3));

Safety

In Rust versions before 1.45.0, some uses of the as operator were not entirely safe. In particular, it was undefined behavior if a truncated floating point value could not fit in the target integer type (#10184);

let x: u8 = (1.04E+17).as_(); // UB

pub fn numcast<D>(self) -> Option<Vec3<D>> where
T: NumCast,
D: NumCast,

Returns a memberwise-converted copy of this vector, using NumCast.

let v = Vec4::new(0_f32, 1., 2., 3.);
let i: Vec4<i32> = v.numcast().unwrap();
assert_eq!(i, Vec4::new(0, 1, 2, 3));

pub fn mul_add<V0: Into<Self>, V1: Into<Self>>(self, mul: V0, add: V1) -> Self where
T: MulAdd<T, T, Output = T>,

Fused multiply-add. Returns self * mul + add, and may be implemented efficiently by the hardware.

The compiler is often able to detect this kind of operation, so generally you don’t need to use it. However, it can make your intent clear.

The name for this method is the one used by the same operation on primitive floating-point types.

let a = Vec4::new(0,1,2,3);
let b = Vec4::new(4,5,6,7);
let c = Vec4::new(8,9,0,1);
assert_eq!(a*b+c, a.mul_add(b, c));

pub fn is_any_negative(&self) -> bool where
T: Signed,

Is any of the elements negative ?

This was intended for checking the validity of extent vectors, but can make sense for other types too.

pub fn are_all_positive(&self) -> bool where
T: Signed,

Are all of the elements positive ?

pub fn min<V0, V1>(a: V0, b: V1) -> Self where
V0: Into<Self>,
V1: Into<Self>,
T: Ord,

Compares elements of a and b, and returns the minimum values into a new vector, using total ordering.

let a = Vec4::new(0,1,2,3);
let b = Vec4::new(3,2,1,0);
let m = Vec4::new(0,1,1,0);
assert_eq!(m, Vec4::min(a, b));

pub fn max<V0, V1>(a: V0, b: V1) -> Self where
V0: Into<Self>,
V1: Into<Self>,
T: Ord,

Compares elements of a and b, and returns the maximum values into a new vector, using total ordering.

let a = Vec4::new(0,1,2,3);
let b = Vec4::new(3,2,1,0);
let m = Vec4::new(3,2,2,3);
assert_eq!(m, Vec4::max(a, b));

pub fn partial_min<V0, V1>(a: V0, b: V1) -> Self where
V0: Into<Self>,
V1: Into<Self>,
T: PartialOrd,

Compares elements of a and b, and returns the minimum values into a new vector, using partial ordering.

let a = Vec4::new(0,1,2,3);
let b = Vec4::new(3,2,1,0);
let m = Vec4::new(0,1,1,0);
assert_eq!(m, Vec4::partial_min(a, b));

pub fn partial_max<V0, V1>(a: V0, b: V1) -> Self where
V0: Into<Self>,
V1: Into<Self>,
T: PartialOrd,

Compares elements of a and b, and returns the maximum values into a new vector, using partial ordering.

let a = Vec4::new(0,1,2,3);
let b = Vec4::new(3,2,1,0);
let m = Vec4::new(3,2,2,3);
assert_eq!(m, Vec4::partial_max(a, b));

pub fn reduce_min(self) -> T where
T: Ord,

Returns the element which has the lowest value in this vector, using total ordering.

assert_eq!(-5, Vec4::new(0, 5, -5, 8).reduce_min());

pub fn reduce_max(self) -> T where
T: Ord,

Returns the element which has the highest value in this vector, using total ordering.

assert_eq!(8, Vec4::new(0, 5, -5, 8).reduce_max());

pub fn reduce_partial_min(self) -> T where
T: PartialOrd,

Returns the element which has the lowest value in this vector, using partial ordering.

assert_eq!(-5_f32, Vec4::new(0_f32, 5., -5., 8.).reduce_partial_min());

pub fn reduce_partial_max(self) -> T where
T: PartialOrd,

Returns the element which has the highest value in this vector, using partial ordering.

assert_eq!(8_f32, Vec4::new(0_f32, 5., -5., 8.).reduce_partial_max());

pub fn reduce_bitand(self) -> T where
T: BitAnd<T, Output = T>,

Returns the result of bitwise-AND (&) on all elements of this vector.

assert_eq!(true,  Vec4::new(true, true, true, true).reduce_bitand());
assert_eq!(false, Vec4::new(true, false, true, true).reduce_bitand());
assert_eq!(false, Vec4::new(true, true, true, false).reduce_bitand());

pub fn reduce_bitor(self) -> T where
T: BitOr<T, Output = T>,

Returns the result of bitwise-OR (|) on all elements of this vector.

assert_eq!(false, Vec4::new(false, false, false, false).reduce_bitor());
assert_eq!(true,  Vec4::new(false, false, true, false).reduce_bitor());

pub fn reduce_bitxor(self) -> T where
T: BitXor<T, Output = T>,

Returns the result of bitwise-XOR (^) on all elements of this vector.

assert_eq!(false, Vec4::new(true, true, true, true).reduce_bitxor());
assert_eq!(true,  Vec4::new(true, false, true, true).reduce_bitxor());

pub fn reduce<F>(self, f: F) -> T where
F: FnMut(T, T) -> T,

Reduces this vector with the given accumulator closure.

pub fn product(self) -> T where
T: Mul<Output = T>,

Returns the product of each of this vector’s elements.

assert_eq!(1*2*3*4, Vec4::new(1, 2, 3, 4).product());

pub fn sum(self) -> T where
T: Add<T, Output = T>,

Returns the sum of each of this vector’s elements.

assert_eq!(1+2+3+4, Vec4::new(1, 2, 3, 4).sum());

pub fn average(self) -> T where
T: Add<T, Output = T> + Div<T, Output = T> + From<u8>,

Returns the average of this vector’s elements.

assert_eq!(2.5_f32, Vec4::new(1_f32, 2., 3., 4.).average());

You should avoid using it on u8 vectors, not only because integer overflows cause panics in debug mode, but also because of integer division, the result may not be the one you expect.

// This causes a panic!
let red = Vec4::new(255u8, 1, 0, 0);
let grey_level = red.average();
assert_eq!(grey_level, 128);

You may want to convert the elements to bigger integers (or floating-point) instead:

let red = Vec4::new(255u8, 1, 128, 128);

let red = red.map(|c| c as u16);
let grey_level = red.average() as u8;
assert_eq!(grey_level, 128);

let red = red.map(|c| c as f32);
let grey_level = red.average().round() as u8;
assert_eq!(grey_level, 128);

pub fn sqrt(self) -> Self where
T: Real,

Returns a new vector which elements are the respective square roots of this vector’s elements.

let v = Vec4::new(1f32, 2f32, 3f32, 4f32);
let s = Vec4::new(1f32, 4f32, 9f32, 16f32);
assert_eq!(v, s.sqrt());

pub fn rsqrt(self) -> Self where
T: Real,

Returns a new vector which elements are the respective reciprocal square roots of this vector’s elements.

let v = Vec4::new(1f32, 0.5f32, 1f32/3f32, 0.25f32);
let s = Vec4::new(1f32, 4f32, 9f32, 16f32);
assert_eq!(v, s.rsqrt());

pub fn recip(self) -> Self where
T: Real,

Returns a new vector which elements are the respective reciprocal of this vector’s elements.

let v = Vec4::new(1f32, 0.5f32, 0.25f32, 0.125f32);
let s = Vec4::new(1f32, 2f32, 4f32, 8f32);
assert_eq!(v, s.recip());
assert_eq!(s, v.recip());

pub fn ceil(self) -> Self where
T: Real,

Returns a new vector which elements are rounded to the nearest greater integer.

let v = Vec4::new(0_f32, 1., 1.8, 3.14);
assert_eq!(v.ceil(), Vec4::new(0f32, 1f32, 2f32, 4f32));

pub fn floor(self) -> Self where
T: Real,

Returns a new vector which elements are rounded down to the nearest lower integer.

let v = Vec4::new(0_f32, 1., 1.8, 3.14);
assert_eq!(v.floor(), Vec4::new(0f32, 1f32, 1f32, 3f32));

pub fn round(self) -> Self where
T: Real,

Returns a new vector which elements are rounded to the nearest integer.

let v = Vec4::new(0_f32, 1., 1.8, 3.14);
assert_eq!(v.round(), Vec4::new(0f32, 1f32, 2f32, 3f32));

pub fn hadd(self, rhs: Self) -> Self where
T: Add<T, Output = T>,

Horizontally adds adjacent pairs of elements in self and rhs into a new vector.

let a = Vec4::new(0, 1, 2, 3);
let b = Vec4::new(4, 5, 6, 7);
let h = Vec4::new(0+1, 2+3, 4+5, 6+7);
assert_eq!(h, a.hadd(b));

pub fn partial_cmpeq<Rhs: AsRef<Self>>(&self, rhs: &Rhs) -> Vec3<bool> where
T: PartialEq,

Compares each element of two vectors with the partial equality test, returning a boolean vector.

let u = Vec4::new(0,2,2,6);
let v = Vec4::new(0,1,2,3);
assert_eq!(u.partial_cmpeq(&v), Vec4::new(true, false, true, false));

pub fn partial_cmpne<Rhs: AsRef<Self>>(&self, rhs: &Rhs) -> Vec3<bool> where
T: PartialEq,

Compares each element of two vectors with the partial not-equal test, returning a boolean vector.

let u = Vec4::new(0,2,2,6);
let v = Vec4::new(0,1,2,3);
assert_eq!(u.partial_cmpne(&v), Vec4::new(false, true, false, true));

pub fn partial_cmpge<Rhs: AsRef<Self>>(&self, rhs: &Rhs) -> Vec3<bool> where
T: PartialOrd,

Compares each element of two vectors with the partial greater-or-equal test, returning a boolean vector.

let u = Vec4::new(0,2,2,2);
let v = Vec4::new(0,1,2,3);
assert_eq!(u.partial_cmpge(&v), Vec4::new(true, true, true, false));

pub fn partial_cmpgt<Rhs: AsRef<Self>>(&self, rhs: &Rhs) -> Vec3<bool> where
T: PartialOrd,

Compares each element of two vectors with the partial greater-than test, returning a boolean vector.

let u = Vec4::new(0,2,2,6);
let v = Vec4::new(0,1,2,3);
assert_eq!(u.partial_cmpgt(&v), Vec4::new(false, true, false, true));

pub fn partial_cmple<Rhs: AsRef<Self>>(&self, rhs: &Rhs) -> Vec3<bool> where
T: PartialOrd,

Compares each element of two vectors with the partial less-or-equal test, returning a boolean vector.

let u = Vec4::new(0,2,2,2);
let v = Vec4::new(0,1,2,3);
assert_eq!(u.partial_cmple(&v), Vec4::new(true, false, true, true));

pub fn partial_cmplt<Rhs: AsRef<Self>>(&self, rhs: &Rhs) -> Vec3<bool> where
T: PartialOrd,

Compares each element of two vectors with the partial less-than test, returning a boolean vector.

let u = Vec4::new(0,2,2,2);
let v = Vec4::new(0,1,2,3);
assert_eq!(u.partial_cmplt(&v), Vec4::new(false, false, false, true));

pub fn cmpeq<Rhs: AsRef<Self>>(&self, rhs: &Rhs) -> Vec3<bool> where
T: Eq,

Compares each element of two vectors with the partial equality test, returning a boolean vector.

let u = Vec4::new(0,2,2,6);
let v = Vec4::new(0,1,2,3);
assert_eq!(u.cmpeq(&v), Vec4::new(true, false, true, false));

pub fn cmpne<Rhs: AsRef<Self>>(&self, rhs: &Rhs) -> Vec3<bool> where
T: Eq,

Compares each element of two vectors with the total not-equal test, returning a boolean vector.

let u = Vec4::new(0,2,2,6);
let v = Vec4::new(0,1,2,3);
assert_eq!(u.cmpne(&v), Vec4::new(false, true, false, true));

pub fn cmpge<Rhs: AsRef<Self>>(&self, rhs: &Rhs) -> Vec3<bool> where
T: Ord,

Compares each element of two vectors with the total greater-or-equal test, returning a boolean vector.

let u = Vec4::new(0,2,2,2);
let v = Vec4::new(0,1,2,3);
assert_eq!(u.cmpge(&v), Vec4::new(true, true, true, false));

pub fn cmpgt<Rhs: AsRef<Self>>(&self, rhs: &Rhs) -> Vec3<bool> where
T: Ord,

Compares each element of two vectors with the total greater-than test, returning a boolean vector.

let u = Vec4::new(0,2,2,6);
let v = Vec4::new(0,1,2,3);
assert_eq!(u.cmpgt(&v), Vec4::new(false, true, false, true));

pub fn cmple<Rhs: AsRef<Self>>(&self, rhs: &Rhs) -> Vec3<bool> where
T: Ord,

Compares each element of two vectors with the total less-or-equal test, returning a boolean vector.

let u = Vec4::new(0,2,2,2);
let v = Vec4::new(0,1,2,3);
assert_eq!(u.cmple(&v), Vec4::new(true, false, true, true));

pub fn cmplt<Rhs: AsRef<Self>>(&self, rhs: &Rhs) -> Vec3<bool> where
T: Ord,

Compares each element of two vectors with the total less-than test, returning a boolean vector.

let u = Vec4::new(0,2,2,2);
let v = Vec4::new(0,1,2,3);
assert_eq!(u.cmplt(&v), Vec4::new(false, false, false, true));

pub fn lerp_unclamped_precise<S: Into<Self>>(
from: Self,
to: Self,
factor: S
) -> Self where
T: Copy + One + Mul<Output = T> + Sub<Output = T> + MulAdd<T, T, Output = T>,

Returns the linear interpolation of from to to with factor unconstrained. See the Lerp trait.

pub fn lerp_unclamped<S: Into<Self>>(from: Self, to: Self, factor: S) -> Self where
T: Copy + Sub<Output = T> + MulAdd<T, T, Output = T>,

Same as lerp_unclamped_precise, implemented as a possibly faster but less precise operation. See the Lerp trait.

pub fn lerp<S: Into<Self> + Clamp + Zero + One>(
from: Self,
to: Self,
factor: S
) -> Self where
T: Copy + Sub<Output = T> + MulAdd<T, T, Output = T>,

Returns the linear interpolation of from to to with factor constrained to be between 0 and 1. See the Lerp trait.

pub fn lerp_precise<S: Into<Self> + Clamp + Zero + One>(
from: Self,
to: Self,
factor: S
) -> Self where
T: Copy + One + Mul<Output = T> + Sub<Output = T> + MulAdd<T, T, Output = T>,

Returns the linear interpolation of from to to with factor constrained to be between 0 and 1. See the Lerp trait.

impl Vec3<bool>

pub fn reduce_and(self) -> bool

Returns the result of logical AND (&&) on all elements of this vector.

assert_eq!(true,  Vec4::new(true, true, true, true).reduce_and());
assert_eq!(false, Vec4::new(true, false, true, true).reduce_and());
assert_eq!(false, Vec4::new(true, true, true, false).reduce_and());

pub fn reduce_or(self) -> bool

Returns the result of logical OR (||) on all elements of this vector.

assert_eq!(false, Vec4::new(false, false, false, false).reduce_or());
assert_eq!(true,  Vec4::new(false, false, true, false).reduce_or());

pub fn reduce_ne(self) -> bool

Reduces this vector using total inequality.

assert_eq!(false, Vec4::new(true, true, true, true).reduce_ne());
assert_eq!(true,  Vec4::new(true, false, true, true).reduce_ne());

impl<T> Vec3<T>

pub fn dot(self, v: Self) -> T where
T: Add<T, Output = T> + Mul<Output = T>,

Dot product between this vector and another.

pub fn magnitude_squared(self) -> T where
T: Copy + Add<T, Output = T> + Mul<Output = T>,

The squared magnitude of a vector is its spatial length, squared. It is slightly cheaper to compute than magnitude because it avoids a square root.

pub fn magnitude(self) -> T where
T: Add<T, Output = T> + Real,

The magnitude of a vector is its spatial length.

pub fn distance_squared(self, v: Self) -> T where
T: Copy + Add<T, Output = T> + Sub<Output = T> + Mul<Output = T>,

Squared distance between two point vectors. It is slightly cheaper to compute than distance because it avoids a square root.

pub fn distance(self, v: Self) -> T where
T: Add<T, Output = T> + Real,

Distance between two point vectors.

pub fn normalized(self) -> Self where
T: Add<T, Output = T> + Real,

Get a copy of this direction vector such that its length equals 1.

pub fn try_normalized<E>(self) -> Option<Self> where
T: RelativeEq<Epsilon = E> + Add<T, Output = T> + Real,
E: Add<Output = E> + Real,

Get a copy of this direction vector such that its length equals 1. If all components approximately zero, None is returned (uses RelativeEq).

pub fn normalize(&mut self) where
T: Add<T, Output = T> + Real,

Divide this vector’s components such that its length equals 1.

pub fn is_normalized<E>(self) -> bool where
T: RelativeEq<Epsilon = E> + Add<T, Output = T> + Real,
E: Real,

Is this vector normalized ? (Uses RelativeEq)

pub fn is_approx_zero<E>(self) -> bool where
T: RelativeEq<Epsilon = E> + Add<T, Output = T> + Real,
E: Real,

Is this vector approximately zero ? (Uses RelativeEq)

pub fn is_magnitude_close_to<E>(self, x: T) -> bool where
T: RelativeEq<Epsilon = E> + Add<T, Output = T> + Real,
E: Real,

Is the magnitude of the vector close to x ? (Uses RelativeEq)

pub fn angle_between(self, v: Self) -> T where
T: Add<T, Output = T> + Real + Clamp,

Get the smallest angle, in radians, between two direction vectors.

pub fn angle_between_degrees(self, v: Self) -> T where
T: Add<T, Output = T> + Real + Clamp,

👎 Deprecated:
Use to_degrees() on the value returned by angle_between() instead

Get the smallest angle, in degrees, between two direction vectors.

pub fn reflected(self, surface_normal: Self) -> Self where
T: Copy + Add<T, Output = T> + Mul<Output = T> + Sub<Output = T> + Add<Output = T>,

The reflection direction for this vector on a surface which normal is given.

pub fn refracted(self, surface_normal: Self, eta: T) -> Self where
T: Real + Add<T, Output = T> + Mul<Output = T>,

The refraction vector for this incident vector, a surface normal and a ratio of indices of refraction (eta).

pub fn face_forward(self, incident: Self, reference: Self) -> Self where
T: Add<T, Output = T> + Mul<Output = T> + Zero + PartialOrd + Neg<Output = T>,

Orients a vector to point away from a surface as defined by its normal.

impl<T> Vec3<T>

pub fn new_point_2d(x: T, y: T) -> Self where
T: One,

Creates a 2D point vector in homogeneous coordinates (sets the last coordinate to 1).

pub fn new_direction_2d(x: T, y: T) -> Self where
T: Zero,

Creates a 2D direction vector in homogeneous coordinates (sets the last coordinate to 0).

pub fn from_point_2d<V: Into<Vec2<T>>>(v: V) -> Self where
T: One,

Turns a 2D vector into a point vector in homogeneous coordinates (sets the last coordinate to 1).

pub fn from_direction_2d<V: Into<Vec2<T>>>(v: V) -> Self where
T: Zero,

Turns a 2D vector into a direction vector in homogeneous coordinates (sets the last coordinate to 0).

pub fn cross(self, b: Self) -> Self where
T: Copy + Mul<Output = T> + Sub<Output = T>,

The cross-product of this vector with another.

On two noncolinear vectors, the result is perpendicular to the plane they define.

The result’s facing direction depends on the handedness of your coordinate system: If we let f be a forward vector and u an up vector, then we have :

Right-handed: f.cross(u) points to the right.

Left-handed: f.cross(u) points to the left.

There’s a trick to remember this which involves your hand: spread your fingers such that your middle finger points upwards and your index finger points forwards, then your thumb points in the direction of f.cross(u).

The following example demonstrates an identity that is easy to remember.

let i = Vec3::<f32>::unit_x();
let j = Vec3::<f32>::unit_y();
let k = Vec3::<f32>::unit_z();
assert_relative_eq!(i.cross(j), k);

pub fn slerp_unclamped(from: Self, to: Self, factor: T) -> Self where
T: Add<T, Output = T> + Real + Clamp + Lerp<T, Output = T>,

Performs spherical linear interpolation between this vector and another, without implicitly constraining factor to be between 0 and 1.

The vectors are not required to be normalized; their length is also linearly interpolated in the process.

let u = Vec3::<f32>::unit_x();
let v = Vec3::<f32>::unit_y() * 2.;
let slerp = Vec3::slerp(u, v, 0.5);
assert_relative_eq!(slerp.magnitude(), 1.5);
assert_relative_eq!(slerp.x, slerp.y);

pub fn slerp(from: Self, to: Self, factor: T) -> Self where
T: Add<T, Output = T> + Real + Clamp + Lerp<T, Output = T>,

Performs spherical linear interpolation between this vector and another, implicitly constraining factor to be between 0 and 1.

The vectors are not required to be normalized; their length is also interpolated in the process.

pub fn unit_x() -> Self where
T: Zero + One,

Get the unit vector which has x set to 1.

pub fn unit_y() -> Self where
T: Zero + One,

Get the unit vector which has y set to 1.

pub fn unit_z() -> Self where
T: Zero + One,

Get the unit vector which has z set to 1.

pub fn left() -> Self where
T: Zero + One + Neg<Output = T>,

👎 Deprecated since 0.14.0:
This function is opinionated about the semantics of X,Y and Z axii, and should not be used. The mapping of axii (X,Y,Z) to perceived directions (e.g right, up, forward) is not universal at all and varies between libraries, graphics APIs, content creation tools and engines. If you want such helper functions, you should write these yourself as part of the package you’re working on, according to what you know about YOUR current coordinate space.

Get the unit vector which has x set to -1.

pub fn right() -> Self where
T: Zero + One,

👎 Deprecated since 0.14.0:
This function is opinionated about the semantics of X,Y and Z axii, and should not be used. The mapping of axii (X,Y,Z) to perceived directions (e.g right, up, forward) is not universal at all and varies between libraries, graphics APIs, content creation tools and engines. If you want such helper functions, you should write these yourself as part of the package you’re working on, according to what you know about YOUR current coordinate space.

Get the unit vector which has x set to 1.

pub fn up() -> Self where
T: Zero + One,

👎 Deprecated since 0.14.0:
This function is opinionated about the semantics of X,Y and Z axii, and should not be used. The mapping of axii (X,Y,Z) to perceived directions (e.g right, up, forward) is not universal at all and varies between libraries, graphics APIs, content creation tools and engines. If you want such helper functions, you should write these yourself as part of the package you’re working on, according to what you know about YOUR current coordinate space.

Get the unit vector which has y set to 1.

pub fn down() -> Self where
T: Zero + One + Neg<Output = T>,

👎 Deprecated since 0.14.0:
This function is opinionated about the semantics of X,Y and Z axii, and should not be used. The mapping of axii (X,Y,Z) to perceived directions (e.g right, up, forward) is not universal at all and varies between libraries, graphics APIs, content creation tools and engines. If you want such helper functions, you should write these yourself as part of the package you’re working on, according to what you know about YOUR current coordinate space.

Get the unit vector which has y set to -1.

pub fn forward_lh() -> Self where
T: Zero + One,

👎 Deprecated since 0.14.0:
This function is opinionated about the semantics of X,Y and Z axii, and should not be used. The mapping of axii (X,Y,Z) to perceived directions (e.g right, up, forward) is not universal at all and varies between libraries, graphics APIs, content creation tools and engines. If you want such helper functions, you should write these yourself as part of the package you’re working on, according to what you know about YOUR current coordinate space.

Get the unit vector which has z set to 1 (“forward” in a left-handed coordinate system).

pub fn forward_rh() -> Self where
T: Zero + One + Neg<Output = T>,

👎 Deprecated since 0.14.0:
This function is opinionated about the semantics of X,Y and Z axii, and should not be used. The mapping of axii (X,Y,Z) to perceived directions (e.g right, up, forward) is not universal at all and varies between libraries, graphics APIs, content creation tools and engines. If you want such helper functions, you should write these yourself as part of the package you’re working on, according to what you know about YOUR current coordinate space.

Get the unit vector which has z set to -1 (“forward” in a right-handed coordinate system).

pub fn back_lh() -> Self where
T: Zero + One + Neg<Output = T>,

👎 Deprecated since 0.14.0:
This function is opinionated about the semantics of X,Y and Z axii, and should not be used. The mapping of axii (X,Y,Z) to perceived directions (e.g right, up, forward) is not universal at all and varies between libraries, graphics APIs, content creation tools and engines. If you want such helper functions, you should write these yourself as part of the package you’re working on, according to what you know about YOUR current coordinate space.

Get the unit vector which has z set to -1 (“back” in a left-handed coordinate system).

pub fn back_rh() -> Self where
T: Zero + One,

👎 Deprecated since 0.14.0:
This function is opinionated about the semantics of X,Y and Z axii, and should not be used. The mapping of axii (X,Y,Z) to perceived directions (e.g right, up, forward) is not universal at all and varies between libraries, graphics APIs, content creation tools and engines. If you want such helper functions, you should write these yourself as part of the package you’re working on, according to what you know about YOUR current coordinate space.

Get the unit vector which has z set to 1 (“back” in a right-handed coordinate system).

impl<T> Vec3<T>

pub fn zyx(self) -> Self

Returns a copy of this vector, with X and Z swapped. This effectively reverses the order of the three elements.

pub fn xy(self) -> Vec2<T>

Same as Vec2::from(self), but shorter.

pub fn with_x(self, x: T) -> Self

Returns a copy of this vector, with a new X value.

pub fn with_y(self, y: T) -> Self

Returns a copy of this vector, with a new Y value.

pub fn with_z(self, z: T) -> Self

Returns a copy of this vector, with a new Z value.

pub fn with_w(self, w: T) -> Vec4<T>

Add a W component to this vector such that it becomes a Vec4.

impl<T> CVec<T>

pub fn into_repr_simd(self) -> Vec3<T>

Converts this vector into its #[repr(simd)] counterpart.